// bedrock-types.ts
export type LatinString = string;
export type ShortArray = Buffer;
export type ShortString = string;
/**
 * UUID is the UUID of the command called. This UUID is a bit odd as it is not specified by the server. It
 * is not clear what exactly this UUID is meant to identify, but it is unique for each command called.
 */
export type uuid = string;
export type byterot = number;
export type restBuffer = Buffer;
export type encapsulated = any;
export type nbt = any;
// ...
export type ItemExtraDataWithBlockingTick = {
  has_nbt: "false" | "true";
  nbt?: {
    version: number;
    nbt: any;
  };
  can_place_on: ShortString[];
  can_destroy: ShortString[];
  blocking_tick: bigint;
};
export type ItemExtraDataWithoutBlockingTick = {
  has_nbt: "false" | "true";
  nbt?: {
    version: number;
    nbt: any;
  };
  can_place_on: ShortString[];
  can_destroy: ShortString[];
};
/**
 * Same as below but without a "networkStackID" boolean
 */
export type ItemLegacy = {
  network_id: number;
  count?: number;
  metadata?: number;
  block_runtime_id?: number;
  extra?:
    | ItemExtraDataWithBlockingTick
    | ItemExtraDataWithoutBlockingTick
    | undefined;
};
// ...
export type RecipeIngredient = {
  type:
    | "invalid"
    | "int_id_meta"
    | "molang"
    | "item_tag"
    | "string_id_meta"
    | "complex_alias";
  network_id?: number;
  metadata?: number | undefined | number;
  expression?: string;
  version?: number;
  tag?: string;
  name?: string;
  count: number;
};
export type PotionTypeRecipes = {
  input_item_id: number;
  input_item_meta: number;
  ingredient_id: number;
  ingredient_meta: number;
  output_item_id: number;
  output_item_meta: number;
}[];
export type PotionContainerChangeRecipes = {
  input_item_id: number;
  ingredient_id: number;
  output_item_id: number;
}[];
export type Recipes = {
  type:
    | "shapeless"
    | "shaped"
    | "furnace"
    | "furnace_with_metadata"
    | "multi"
    | "shulker_box"
    | "shapeless_chemistry"
    | "shaped_chemistry"
    | "smithing_transform"
    | "smithing_trim";
  recipe:
    | {
        recipe_id: LatinString;
        input: RecipeIngredient[];
        output: ItemLegacy[];
        uuid: string;
        block: string;
        priority: number;
        unlocking_requirement: RecipeUnlockingRequirement;
        network_id: number;
      }
    | {
        recipe_id: LatinString;
        width: number;
        height: number;
        input: RecipeIngredient[][];
        output: ItemLegacy[];
        uuid: string;
        block: string;
        priority: number;
        assume_symmetry: boolean;
        unlocking_requirement: RecipeUnlockingRequirement;
        network_id: number;
      }
    | {
        input_id: number;
        output: ItemLegacy;
        block: string;
      }
    | {
        input_id: number;
        input_meta: number;
        output: ItemLegacy;
        block: string;
      }
    | {
        uuid: string;
        network_id: number;
      }
    | {
        recipe_id: LatinString;
        template: RecipeIngredient;
        base: RecipeIngredient;
        addition: RecipeIngredient;
        result: ItemLegacy;
        tag: string;
        network_id: number;
      }
    | {
        recipe_id: LatinString;
        template: RecipeIngredient;
        input: RecipeIngredient;
        addition: RecipeIngredient;
        block: string;
        network_id: number;
      };
}[];
export type RecipeUnlockingRequirement = {
  context:
    | "none"
    | "always_unlocked"
    | "player_in_water"
    | "player_has_many_items";
  ingredients?: RecipeIngredient[];
};